// src/api/models/questionGenerationRequest.ts
/**
 * Generated by orval v8.0.3 🍺
 * Do not edit manually.
 * EdCraft Backend API
 * API for EdCraft Backend
 * OpenAPI spec version: 0.1.0
 */
import type { ExecutionSpec } from './executionSpec';
import type { GenerationOptions } from './generationOptions';
import type { QuestionSpec } from './questionSpec';

/**
 * Request to generate a question based on form selections.
 */
export interface QuestionGenerationRequest {
  /** Original algorithm source code */
  code: string;
  /** Specifications for the question to be generated */
  question_spec: QuestionSpec;
  /** Specifications for code execution */
  execution_spec: ExecutionSpec;
  /** Options for question generation */
  generation_options: GenerationOptions;
}